+++ lib/app/ui/widgets/snackbar.dart
@@ -1,0 +1,34 @@
+import 'package:flutter/material.dart';
+
+class Snackbar extends StatelessWidget {
+  const Snackbar({super.key});
+
+  static error({required BuildContext context, required String title}) {
+    ScaffoldMessenger.of(context).showSnackBar(
+      SnackBar(
+        margin: const EdgeInsets.all(10),
+        behavior: SnackBarBehavior.floating,
+        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(5)),
+        backgroundColor: Colors.red,
+        content: Text(title),
+      ),
+    );
+  }
+
+  static success({required BuildContext context, required String title}) {
+    ScaffoldMessenger.of(context).showSnackBar(
+      SnackBar(
+        margin: const EdgeInsets.all(10),
+        behavior: SnackBarBehavior.floating,
+        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(5)),
+        backgroundColor: Colors.green,
+        content: Text(title),
+      ),
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}